great changes were made

+++ app/routers/user.py
@@ -2,23 +2,22 @@ from typing import List, Optional
 from fastapi import FastAPI, Response, status, HTTPException, Depends, APIRouter
 from sqlalchemy.orm import Session
 from ..config.database import get_db
-from app import models, schemas, utils
+from app import models, schemas, utils, oauth2
 
 router = APIRouter(prefix='/users', tags=['Users'])
 
 
 @router.get('/', response_model=List[schemas.UserRes])
-def get_users(db: Session = Depends(get_db), limit: int = 0, skip: int = 0, search: Optional[str] = ""):
-    users = db.query(models.User).all()  # gets all available results
-    # users = db.query(models.User).limit(limit).all()       #limits to x number of results
-    # users = db.query(models.User).offset(skip).all()       #skips x number of results
-    # users = db.query(models.User).filter(
-    #     models.User.name.contains(search)).limit(limit).offset(skip).all()
+def get_users(db: Session = Depends(get_db), current_user: dict = Depends(oauth2.get_current_user), limit: int = 0, offset: int = 0, search: Optional[str] = ""):
+    if current_user.is_super_admin != True:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,
+                            detail=f"Not Authorized to perform requested action")
+    users = db.query(models.User).all()
     return users
 
 
 @router.get('/{id}', response_model=schemas.UserRes)
-def get_user(id: int, db: Session = Depends(get_db)):
+def get_user(id: int, db: Session = Depends(get_db), current_user: dict = Depends(oauth2.get_current_user)):
     user = db.query(models.User).filter(models.User.id == id).first()
     if not user:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
@@ -27,8 +26,10 @@ def get_user(id: int, db: Session = Depends(get_db)):
 
 
 @router.post('/', status_code=status.HTTP_201_CREATED, response_model=schemas.UserRes)
-def create_users(user: schemas.UserCreate, db: Session = Depends(get_db)):
-
+def create_users(user: schemas.UserCreate, db: Session = Depends(get_db), current_user: dict = Depends(oauth2.get_current_user)):
+    if not current_user.is_super_admin and not current_user.is_admin:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                            detail=f"Not Authorized to perform requested action")
     hashed_pass = utils.hash(user.password)
     user.password = hashed_pass
     new_user = models.User(**user.dict())
@@ -40,24 +41,32 @@ def create_users(user: schemas.UserCreate, db: Session = Depends(get_db)):
 
 
 @router.delete('/{id}', status_code=status.HTTP_204_NO_CONTENT)
-def delete_user(id: int, db: Session = Depends(get_db)):
+def delete_user(id: int, db: Session = Depends(get_db), current_user: dict = Depends(oauth2.get_current_user)):
     user_query = db.query(models.User).filter(models.User.id == id)
     user = user_query.first()
     if user == None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                             detail=f"user with id: {id} was not found")
+    if not current_user.is_super_admin and not current_user.is_admin:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,
+                            detail=f"Not Authorized to perform requested action")
+
     user_query.delete(synchronize_session=False)
     db.commit()
     return Response(status_code=status.HTTP_204_NO_CONTENT)
 
 
-@router.put('/drugs/{id}')
-def update_user(id: int, updated_user: schemas.UserCreate, db: Session = Depends(get_db)):
+@router.put('/{id}')
+def update_user(id: int, updated_user: schemas.UserCreate, db: Session = Depends(get_db), current_user: dict = Depends(oauth2.get_current_user)):
     user_query = db.query(models.User).filter(models.User.id == id)
     user = user_query.first()
     if user == None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
                             detail=f"user with id: {id} was not found")
+    if not current_user.is_super_admin and not current_user.is_admin:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,
+                            detail=f"Not Authorized to perform requested action")
+
     user_query.update(updated_user.dict(), synchronize_session=False)
     db.commit()
     return user
